backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
numpy==1.26.4
scikit-learn==1.5.2
google-generativeai==0.3.2
python-dotenv==1.0.1
websockets==12.0
pydantic==2.6.3
